Fix bug in type-spec->type
Turns out that make-instance also works for structs. type-spec->type
uses make-instance on the name you gave to get the type
instance. However we have shadowing for the very reason that a given
name may represent a lisp type.

In these case the bug caused an instance of the lisp type to be made and
returned.