make #'type-specp shadow aware
turns out that #'type-specp checks a typespec is valid by trying to
tyurn it into a type, if it failed it returned nil. However it wasnt
looking at shadowed symbols so it was having issues with cepl structs.

To fix this it now just calls #'type-spec->type and catches any
unknown-type-spec error. If it gets that error it return nil, else it
returns t.