struct & labels fixes
Two bugs down, one a real nasty:

The first is we now make sure #'type->type-spec returns a the 'true'
type of a struct if relevent. Fake structs are only used in the case of
struct vertex arrays and are created by the compiler on demand.
We also make sure v-true-type works for 'spec' types

Now for the nasty one. This code

    (varjo::defshader bam ()
      (let ((x (v! 0 0 0)))
	(labels ((wam ((_0 :int) (x :vec3))
		   x))
	  (v! 0 0 0 0))))

was cause real issues but ONLY if there were two arguments in wam and
the second one shadowed an earlier arg.

Turns out that #'merge-env had a dumb bug that caused invalid
environments when names collided and environment being merged had
multiple entries for that collided-var.

Anyway that took too long to fix!