Start adding flow-ids
We need some basic flow control. My current id is to give values an id
when they are created and then to let the id propagate through forms as
the code is compiled. It will allow us to tell where values go and thus
what values are transformed by what functions.

The driver for this is the 'spaces' feature for cepl. It wants to know
every place a space is used so it can upload only the needed transforms.

Enough for tonight, ciao