Various bugfixes
Main reason for commit is that im about to kill some code. It was wrong,

The code was meant to find variables by flow-id so I could update
them (for setf) it's wrong because we need to update the binding not the
flow-id for example:

  (let ((a 1)
        (b a)
        (c a))
    (setf b 10))

if we update by flow-id it would be all of these, which is crap because
it's b we are updating.

So yeah, commiting, killing that code and trying something else