remove the crap, found problems with places
Oh man Ιknew this would come back eventually. Places. I hacked em in to
varjo and they are wrong in a bunch of ways. I'll spare you the details
but I'm gonna pull them out and reimplement it all.

The first step though was implementing 'setq' which is done. Now I need
to decide how Ι will handle forms (aref etc), and what kinds of
limitations I will put on this.

There is no point in this being fully general as there are only arrays,
vars and structs on the gpu.