I'm about to totally change how environment works
Up until now we have cloned the environment all over the place to
protect from side effects. We have also had a single environment
represent the entire state at that time.

Now we have switched to a tree of environments. We still want to be
able to mutate however so we will be forking the env tree often.
This means all functions need to return the environment.
The functions that mutated directly will now fork. So all in all its
just more functional than before.

Ho hum this is chaos