extend ast-node
hacked in support for both kinds of flow origin and also parent node

parent node is needed so you can browse up to the nearest
`whatever`. For example the nearest let scope, or maybe even just get to
the surrounding form.

the `origin`s are terribly named. I need to come up with some decent
terminology for this, as currently it sucks.