rolling-translate can take custom compile function
Internally #'rolling-translate uses #'translate to compile the lisp
code, we now allow you to pass into your own translate function.

An example of a use for this is that cepl uses varjo in a multi-pass
fashion. It has #'v-translate which calls varjo's #'translate on all
it's passes until no more changes are detected. By passing #'v-translate
into #'rolling-translate cepl gets multi-pass compilation even inside
the rolling-translate process, which is kinda cool!

This also means that varjo doesnt have to worry about this stuff yet. It
is very likely that varjo will eventually provide it's own customizable
pass system, but for now this keeps the complexity where it belongs.