add ast-deep-replace to varjo
This allows a deep search and modify of the ast graph. Once a change has
made it early-outs of the entire walk and returns the modfied code.

One important thing to note is..hmm how to best explain it. Lets say we
are matching on the number 1.

```
     ((((a ((1))
	 b ((1))
	 c (((1 1)))))))
```

In this case it will match BOTH the nodes under `c` and NEITHER of those
under `a` or `b` (even though they were walked first)

This is because it looks for the deepest match. If it is undecidable it
brings all deepest nodes.