Fix #17 - loop flow-id bug
Ok so this was a two part bug. First lets look at the code that
triggered it:

```
(defshader test-pipeline ()
  (labels ((inner-func ((d1 :vec2) (d2 :vec2))
	     (let* ((result (v! 0 0)))
	       (%if (< (x d1) (x d2))
		    (setf result d1)
		    (setf result d2))
	       result)))
    (let* ((d (v! 0 0)))
      (for (i 0) (< i 5) (setf i (+ i 1))
	   (setf d (inner-func d (v! 1.0 1.0)))))
    (v! 0 0 0 0)))
```

Cool so the tricky bit is that the second arg to inner-func is generated
inside the loop. This means the function won't reach a fixpoint.

One straightforward but hard to find bug was that #'id= was broken. It
would sort the raw-ids correctly but then it would call #'equal on the
sort sorted lists of raw-ids. Instead it should have been comparing the
value contained in the raw-ids.

The other thing I did was to add checkpointing of flow-ids. This means
before analysing the loop a checkpoint is generated, then 1 pass of the
loop is analyzed, after this it resets to the checkpoint meaning the
same flow-ids are allocated again. this stops new flow-ids breaking the
resolve process.
If a fixpoint has been reached we dont reset the checkpoint but instead
continue.