Move iuniforms (stemcells) bool to env object
The way we handle the context has always been a bit shitty. The idea of
sticking it in a list is fine for the lambda list but sucks for actually
doing checks against.

We need to make a proper fix where we add more slots to the environment
and parse the context values properly..but that's a task for another
day. Actually I'll go raise and issue for that now. Ok done that :)

The first step is finished though I moved the :iuniform arg into it's
own slot. It's used to say where the stage can use implicit uniforms and
it's ok for this value to be different for different stages.