Allow name collision for safe cases
Sometimes we choose to compile using functions & variables from all
stages (for example this is done for code sanity checking by CEPL) in
those cases we will get clashing definitions for the glsl variables.

For example gl-clip-distance is in vertex, geometry & fragment.

We sanity check those kind of collision and allow them in veyr protected
cases