added a predicate for testing if the ast converges
When you get an ast from the compile result it may be slightly different
that the code you compiled. This is because macros will have expanded
and special-forms may have modified things.

This function:
- takes the ast
- generates new lisp code from the ast
- compiles that lisp code
- gets the ast of the new compile result
- generates lisp code from this new ast
- compares the lisp code from the two ast's and returns t if equal.

We allow this recompile chain to happen up to 'max-depth' times in order
maximize the chance of the code converging.