Fix bug where user/local func causes flow-id error
In cases where a local or user defined function created a new flow-id
the flow analyzer could error as it couldnt find this new flow-id in the
current flow-ids of the environment. Now instead of erroring we say that
the call site is the origin of the flow-id.

One other option would be for 'labels' to add the flow-ids of it's
internal ast-nodes to the env..I'm not sure this makes sense yet though