remove to-top in favor of adding str to func cache
It's turns out that to-top was only used for functions. We also found
out that #'search-for-flow-id-fixpoint would break code as multiple
iterations would result in varjo mistakenly thinking it has already seen
a function and deduping it.

We really need to fix (or remove) the flow system, however this is still
fix avoids the issue without breaking other code. So for now it wins