add restart for when exact function not known
When we test-compile a function we don't know the exact functions that
will be passed into the stage. This would usually result in a crash.

We now make use of a restart to continue the compile process, this will
creates a dummy function with the arg & returns types from the
spec. This means that type checking is still valid.