emit funcall ast node when cant establish func
When we compile a funcall we work out the function and then compile it
as a regular function call.

Sometimes we don't yet know the function object (like when compiling a
single function with function args). In those cases we create a dummy
function and compile that.

The issue was that it meant that the ast contained a regular call to a
function with a name like 'dummy-function'. If you then try and
recompile that AST you are going to get a 'dummy-function not found'
error.

To avoid this we create a new kind of AST node: the funcall node. This
means we can regenerate valid lisp code from the AST.