Start adding function-set support
The idea is that in places where the user types something like #'sin we
make the ctv be the set of all accessible overloads of 'sin' at that
point. We can then pass this set of functions around and at 'funcall'
forms we can use the standard function scoring mechanism to choose the
correct one to use.

We also change 'funcall' so the ast generated is of the form:

    (funcall x ,@y)

the reason for this is that if we expand to the exact function-name then
we can often run afoul of scoping issues. By doing this we guartentee
the ast can stabilize