pass allowed-outer-vars in environement
When compiling functiosn that take functions we need to inject a
'labels' form. This works but has been resulting in unwanted variable
capture. However we couldnt just switch to labels-no-implicit as it did
allow the capture of the function var we wanted to capture.

We now allow labels-no-implicit to take an optional list of exceptions,
vars that can be captured. We pass this info down in the
environment.

This allowed us to fix test build-32, which has been hanging around for
a while