fix labels, flet, %make-function environment bug
Ok, so the issue was that we use the environment we had made for the
function as the environment we added the function to.

This was also the environment we compiled the body of the
flet/labels/etc in, so as you can imagine this caused some issues. The
symptom that I originally noticed was that the function-scope id was too
high.

The 'funny' thing is how long we got away with this for. The reason was
that (except for labels-no-implicit) the functions could read data from
the surrounding environment, so it just worked :|

Ah well, all the tests pass now. I'm off to test some of my larger
projects and see how they behave with this change