good progress on external-function call rewrite
The old mechanism of expanding function into labels-no-implicit scopes
was unreliable. I'm replacing this with code that compiles them with the
base-env and also reuses #'build-function & #'compile-function-call to
make sure we handle everything we want fluidly.

I'm not really happy with how we use the dedup pool for the source of
glsl for the compiled functions. I want to move this somewhere better. I
also need to think about the dedup pool in general, given that now we
don't just inject labels-no-implicit forms (which really pushed the need
for the dedup).

That's all for now. Peace