add compiled-function-result
I have moved some of the logic from compile-with-external-func-in-scope
to build-external-function as that seemed to be a better place for it. I
have started saving the compiled results to base-env's cache.

Currently function declarations arent emitted but will be in the next
step where we will remove the current dedup system and use the
'compiled-function-result's to generate the final glsl result