finally remove v-t-type
It's been moot for a while but I have finally done the cleanup to get
rid of the damn thing.

Originally it was used for the glsl spec-types, a hack to try and avoid
expanding all the types. This didnt work well and the glsl-spec project
was the result. Now that has been solid for a while I'm happy to remove
these, especially as it had got to the point of me not being able to
justify why void and none were not of type v-type