Rewrite 'if'
Make the standard `if` much more flexible by using the `or` type.
We can now return a single type if both expression's types match or an
`or` type if they don't.

Even though we can't yet represent the 'or' in glsl, by allowing the
compilation to continue we can give better error messages and we also
ignore non matching types when the result isnt used.

For example:

```
(glsl-code
  (varjo.tests::compile-vert () :450 nil
    (if (< 1 2)
        1
        4s0)
    (v! 0 0 0 0)))
```

compiles to

```

void main() {
    if ((1 < 2))
    {
        1;
    }
    else
    {
        4.0f;
    }
    gl_Position = vec4(float(0),float(0),float(0),float(0));
}
```

And this is fine as even though the 2 clauses of the `if` differ in
type (and we returned an (or :int :float) type) the user didnt try to
pass the value to a function or return it.