v-compile-time-value -> v-unrepresentable-value
In the future I want any type to be able to have ctv information. For
example an :int where we know the value is between 0 and 10.

This means v-compile-time-value is a bad name. What it really is is a
value that is unrepresentable as anything else. This is why it has to be
spliced in with function injection.