add the ephemeral-type
Ephemeral gets the same glsl stripping as unrepresentable types but no
function splicing.

I want to rename unrepresentable types.

Also removed the pass-by-ref type for now. It was confusing things, this
should likely be a property of the type

hmm, now I think about it, we could remove unrepresentable and add a
function splicing field to ephemeral types (or the other way
around). I'll ponder this some more