Fix 2 first class func bugs
The first bug what that secondary-score could be nil and we were
ignoring matches if that was the case. A secondary score is only
relevent if the function already matches so instead of null we now
return most-positive-fixnum

The next was that cast-code couldnt be specialized on the varjo type of
the obj. To that end we make cast-code a function that dispatches to
cast-code-inner (which is a method) with the code-objects type as an
additional argument.

This last one let us make special case of cast-code-inner for
v-any-one-of types.