Added typed-var-macro
This is a new kind of macro I'm playing with. It will be the companion
to compiler-macros (which will be typed in Varjo).

It lets you declare a macro that will be called whenever a variable of a
specific type is accessed. This allows you to do symbol-macro like
expansion based on the type of a variable.

I'm using this to hook into when spatial vectors are accessed so I can
see if they have entered a new spacial scope.