Callbacks for adding metadata on type construction
This allows us to infer metadata for a deftype'd type whenever one is
compiled. First you now have to declare the metadata kinds that are
valid for a type in deftype (let's call an example type FOO). Next,
whenever FOO is the type of a compiled form and the metadata is not yet
set, the infer calling is called and you are given an opportunity to
return the required metadata (or nil to ignore it).