Use combine-metadata on every set of metadata
We use combine-metadata as the mechanism to resolve all metadata
updates. Combining with nil is a no-op but anything else is handed over
to the user's own specialization of combine-metadata.

I think I'm going to change the calling code so that it can accept a
metadata object or the args to a make-instance of the metadata-type
this makes it the same and the infer mechanism