WIP: first sign of progress in new 'out' logic
Mid way through some rewriting but this is a good spot for a checkpoint.

We use glsl-expr in 'return' to replace the old %out stuff. We are
moving to deterministic names for the out args. Currently there may
be clashes between stages but we can simply add the stage name to the
out-var in future to resolve this.