WIP: Small pile of cleanup things
Made the inteface block stuff a bit better. Names are matching, glsl
looks saner.

Next task is making sure 'return' is emitted in glsl in the right
places amd that we can return void from the implicit return, hmm maybe
we should be smarter about that, like if there is a return-set and we
could tell that it was from a form from the tail, then we could ignore
the mismatched out-types.. maybe if we add a 'merged-return-set' type we
could handle this. Anyway, home first, then look at this.