WIP: Remove redundant 'returns' from code object
Well this took some fuckery :p

As we have return-set we dont need 'returns'. I didnt even notice the
duplication when I started.

I was seeing issues arrising from 'type' and 'multi-vals' in the
code-obj being incorrect. The reason is that those two are for
communication up to the stack _not_ really for return info. It's a
subtle difference but important, return-set is the only valid
communicator of the return vals of a function.

I still think that code-type should be merged with multi-vals. This
would mean we have only code-type and return-set, and that captures what
we need nicely