emit glsl for ephemerals, but hidden behind flag in current-line
This is handy for the odd occasion where we do want to do something
with the string but for all other purposes the type must be
ephemeral.

The case for this is the block-array, we need to make a glsl string
for aref, but we dont want anything else to use it.