Remove multi-stage and add proper test compilation
The multi stage type was doomed to fail once variables came into the
picture. We now add the test-translate/-raising functions which will
take a stage and try to compile it as many kinds of stage.

It then collates the results and/or errors and optionally can raise
and error summarising the problems across the attempted compilations.