More logic around the determining purity
We like being able to remove functions which have no impact on the
code. We did this by not emitting code for any user defined function
which had void return (as local funcs only allow local
mutation). However this does not take into account that some glsl
functions do cause global mutation (atomics, emit-vertex, etc).

We now use the updated glsl-spec (which include purity information)
and propagate info on purity as we compile. We then use this info when
we are deciding whether to emit the glsl.

Right now the code will probably mark some things are impure that are
pure so please file bugs as you find these cases.