Facility for defining own constructors for types :)
When a list typespec is found that is not an array spec and whose
first element is the name of a subclass of v-type we allocate an
instance of it and pass it to #'v-make-type along with the #'rest of
the spec.

There the author of the type is meant to call #'initialize-instance
and return the result. Clearly they are also allowed to do whatever
they like with the args from the typespec.

This finally makes the specs extensible in a seemingly sane way.

Using it we now handle array, 'or' & functions types in a consistant
way, and also add support for constructing ephemeral arrays