Add support for variables not yet in glsl-spec project
glsl-spec is good but not complete, we want to try out some things
before we push them upstream so we add support for that with
the *definitions-missing-from-glsl-spec* variable.

We also add support for the "{typeName" pattern which means the type
is a member of an array'd interface block. We represent that
internally with an ephemeral-array.