Tests & fixes for returns
Things were getting a little out of hand, it was just too easy to
'fix' something in the `return` mechanism and break something else.

A few tests nailed down the problem cases and we are now back in
business.

Also something funky is going on. Not happy with it, why hasnt this
come up sooner? why do we sometimes see returns sets containing just a
void val.. more testing needed to find this