Added scoped metadata
Let's you define and use declarations that arent bound to variables,
but instead to the current scope.

We then attach top-level meta from fucntion body to function
definition & emit geometry outs.

Finally got something that feels sensible for geometry shaders, we can
make the 'out' declaration. A nice unexpected side effect of using
declare for this is that the function is still valid for other stages,
as the compiler is allowed to ignore whatever it likes.

The current mechanism around this is still a little hacky, but it's
good enough for testing. We can do more soon.