Implement #'emit-data
It works! So now we can add per vertex data to each vertex from a
geometry stage.

To do this we copied return-set related code add similar for the newly
added emit-sets. We allow these to flow over function boundaries so we
are sure that every emit-set makes it to the top and is checked.

Rock on!