Add glsl-to-compile-result & fixes for glsl geometry shaders
I've moved the logic for compiling glsl stages from CEPL to here. It
makes way more sense as part of Varjo, hopefully I can then write more
tests for this and keep it in step with the rest of Varjo's
development.

Also in this commit are fixes to get glsl geom stages working. I had
to give up the idea of swapping out the `in` block array access for
now. Regex is out of the question without invoking tony the pony and
Shimera's glsl parser isnt yet ready to handle all valid glsl.