Get Tessellation working (for real :p) mainly dealing with impl shit
OK so we have had to dance a merry dance to get here, but it is kinda
working. We ran into an issue where my linux machine didnt like
tessellation in/out blocks without instance names, this then caused
another problem as tessellation outs/in have to be array'd, so
suddenly every fucking thing need ephemeral arrays. What a pain in the
gooch.

Here are some changes:

- emit only used for geometry now, everything else uses return
- add the tessellation-stage meta stage class, is not itself a stage
- tessellation-evaluation correctly writes first return to gl-position
- move #'process-output-primitive early in compilation so we can use
  the vertex count
- add #'make-out-set arrays the raw-out-set when needed
- more stages using ephemeral block arrays
- make the length of the 'in' blocks for tessellation stages be
  gl_MaxPatchVertices (fixes an error on my machine)
- nicer formatting for 'multiple values' glsl names
- ensure no duplication of multiple value base name (stops clashing
  var definitions with multple returns in same scope)
- make #'postfix-glsl-index as we had duplicate string formatting in a
  few places and it would be too easy to break
- remove unused #'fmt
- move 'gl_InvocationID' indexing into nth-return-name