We were trying to handle value-safe in a form it wasnt used
The bug manifest as a value-safe-wasnt-safe error. It would happen
when you called a function on two functions returning multiple
values.

The reason is that we have a special form called 'values-safe' which
allows multiple-return values to flow through a regular function call
form. This isnt usually allowed but 'values-safe' allows it for 1
form. The problem was that we were calling the function to handle that
case even when 'values-safe' hadnt been used.

Blimey