Propagate environment through function argument compilation
Arguments are compiled before the function call, and the forms in the
arg positions have the potential to affect local state. For example:

   (let (((i :int)))
     (v! 0 (setq i 1))
     i)

This means we have to propagate the env though the args and into the
function call itself. The trick with merging the arg env with the
original env is so that things like knowlege of 'values' etc are maintained.