Force users to return (values) from compute stages
Originally I had thought that given then it was the only valid return
for a compute stage I may as well do it for the user. However a
gpu-func in cepl can technically be reused for any stage so I've
decided to require the user writes it.

I needed to update the 'test compile' code to respect this too

This also should fix an ast gen bug.