parse qualifiers sooner and make sure code uses qualifier=
A lot of code was assuming a qualifier was a keyword. However once
they are parsed they are an object, this means we needed to revisit
all the code that deals with qualifiers and make sure that it was
testing correctly.

The main helper for this was making the predicate #'qualifier= . This
let use compare qualifier objects and keywords.