remove qualifiers from shader-variable. Use qualifiers from type
Type already had this info so the duplication was unneccesary and
confusing. The changes in the code are mostly just updates to extract
the data from the type rather than from the variable object.

We strip the qualifiers from the vars when making the values as
otherwise the qualifiers can flow to the outvars without qualifiying
using `values`

We had to tweak our ssbo detection in assignments as that relied on
qualifiers before

One other change of note however was that the block-array's copy-type
was not copying the flow-ids, which got caught during this run of tests.