Dont emit function code when compiler macro expanded instead
We now keep track of how many times a compiled function was
called. This is important as a function's compiler macro could replace
the call with something else. In that case we don't want to add the
glsl for the function definition (or signature) but we do want to know
that the function was used. Whether a function has been used is
important to projects like CEPL.