Change mosts test to complile using 410 (for macOS)
The reason is that, if we load cepl.tests after varjo.tests we get
automatic compilation testing on the gpu. However 4.1 is the highest
GL version apple supports so we use that.

cepl.tests is going to look at the compiled-stage's context to see if
the GL context it has is at least the version the test is requiring