More work on used-type propegation, copy-compiled doing less magic
copy-compiled was adding the type-set to the used-types. This sucked
as it means copying an instance of 'compiled could end up with
additional used-types. We remove that logic and do more work in the
places where make/copy-compiled are used.

We also work a bit harder to ensure that function objects store all
the used types from their definition as we rely on all of them getting
added when a call to the function is compiled.

With these changes we get even fewer unused struct definitions in the
glsl:

```
	TESTS> (glsl-code
			(compile-vert ((vert :vec4)) :410 nil
			  (flet ((foo ((x pos-col))
					   (pos x)))
				(let ((a #'(foo pos-col)))
				  (v! 1 2 3 4)))))
	"// vertex-stage
	#version 410

	layout(location = 0)  in vec4 VERT;

	void main()
	{
		gl_Position = vec4(float(1),float(2),float(3),float(4));
		return;
	}
	"
```