Fix return-15, the last of the failing return tests
This is an interesting case, it is possible for the user to write a
function where they always discard. Whilst strage I don't think is
illegal so I dont want to throw an error here. We have to make a valid
function though so it pretty much has to be a void.

It feels super nasty as nil usually isnt a valid return-set and we
just say 'ah it's fine just use void' but I think the logic is ok for
doing this here.