Refactor how we store & generate glsl-code
Store lines of glsl as objects with a string and an indentation. This
makes indentation cheap as we just increment the value.

lines are stored in chunks which allow indentation too. With a few
helper functions we have something that is much nicer to read
and (hopefully) more efficient

The glsl generation stuff now has its own dir