generate a glsl-alphanumeric-p function
The scanning of that string was super dumb (not a surprise other than
how long we left it there) but we couldnt just replace it with
alphanumeric-p becuase that would return true for chars like λ.

We also couldnt just assume ascii codes as charset is not specified by
the spec (luckily). So we generate our own function that checks the
charcode sits withing certain ranges that we generate at
macroexpansion time.

Totally works and gives us a millisecond or so back :)