first hash at extensible sequences, getting closer
Two big issues. making traits unreps is not ideal, I'd rather inline
the func (like an unrep) and replace the types with the one from the
args.

Also deduplication needs to be better as two calls to mapseq result in
two definitions in the glsl which is obviously bollocks :D