our traits are still to basic to encode map & reduce properly
Functions taking traits need to be able to have args that are typed
based on the trait-args. for example reduce wants an initial value,
this should be the 'elem-type' of the sequence argument, but we cant
do this yet.

Also Im using compiler macros to inline map & reduce as currently the
monomorphised version of the functions is specific to the array length
so mapping over differently lengthed arrays results in different
functions in the glsl