Tentative avoiding a var declaration when val is just another var
Shitty constant folding :D This would be so easy in a proper compile,
but alas, this is not that compiler.

Seriously though I would do make the compiler totally differently
today but Im not scrapping this one as it still is useful; even with
it's madness. Varjo has been a great vehicle to learn about compilers
and ne day it'll likely get a rewrite, but I'm going to practice some
more in other projects first.