Fix spurious glsl func in glsl + another check for recursive inline
I can't think of a case where it makes sense to have an inline of a
function that takes unreps without any of the args being moved to the
surrounding scope. If there are no args changes then the inlining is
going to be recursive (which is exactly what we saw a couple of
commits back). Now we have a cheap check and a big old error if it
goes wrong.

We also track compiler-macro expansions so we know not to increment
the call count of a function if it expanded into something else. I've
decided that this is valid for any form of expansion (like inlining
for unreps) so we have renamed the var and made sure it propagates in
the right places.