Refactor search for flow-id fixpoint
This code has filled me with dread for ages because it's crufty as
fuck and because I know I need to address the whole base-env mutation
thing.

In the process of trying to understand it I saw a pile of things that
made the code harder to read so I've refactored and documented as I
went. All tests still pass.