dont export the shift operators
This compromises the tradeoff we made with not exporting
glsl-symbol.operators.

Below is the reason we gave for not reexporting from the original
commit.

 - - - - - - - - - - - - - - - - - - - -

Those symbols are too likely to clash with something else. The most
important ones are in CL anyway and the rest can be used by :use'ing
glsl-symbol.operators yourself.

It's a more reasonable balance.

This came up as 5am clashed on '!'