Change how function stripping works
This was done to fix a bug where passing ssbos across functions caused
spurious functions in the glsl.

The way he fix it is to accumlate compiled function results in the
'compiled' object. Then we can walk this info recursively from the
main function and we get only the funcs that were actually called.

Also done:

- move xxbo-p out to it's own function
- remove all the old call-count code (as it was incorrect)