flesh out the complex api
Now support single-complex and double-complex types. The types complex
and v-complex are aliases for single-complex.

Support for double-complex is limited by glsl not providing trig
functions for doubles.

sqrt is based on sbcl's version but without any of the overflow
handling or nan safety (fits more with glsl this way)

A bunch of our trig functions are based off the mathematical form of
the problem and may have unknown precision and performance issues.
SBCL's version take much more care about overflow, nan etc