when varjo cant find a var check if its defined for other stages
This is useful when trying to use gl vars (e.g. gl-invocation-id)
which are only defined for certain stages. Previously the error would
just say that the var wasnt defined which, while true, isnt the most
useful for diagnosing the issue.

Nowe we get an additional note that tells you which stages it is
defined for.