add &inline-form arg to with-slots
So the issue is that when with-slots binds a value to a temp variable
it's is copying the value. This means atomic-add is confused as it can
only work against things like block memory.

To this end, we added the keyar &inline-form to with-slots. This
allows you to declare that the form must be inlined in the slot
expansions. Our only restriction is that the form must be pure.